Add files via upload
Hi, So the question is to add two numbers as a list. if we want to add them to a single linked list. We can go to our basic mathematics so add the two values and track the carry throughout the process. Initialize a temp list and their pointer. it will move to store the sum of the two linked lists in the temp list. so start iterating both the list and add and check carry if it's non-zero at the end of the loop check if carry is a non-zero value if so we can add a new node for the carry at the end.

Time Complexity:- O(n)
Space Complexity:- O(1)