Add files via upload
Hi, so the question is the longest substring without repeating characters. for this question, we have to understand the concept. 
Neive approach or simplest approach is to generate the all the substring and check for if it is non repeating substring so the,
Time Complexity:- O(n^3)
Space Complexity:- O(1)

So, The optimized approach uses two pointers left and right, and takes a set to store the characters in it. so in this optimized approach we use left and right pointers to 0 and move the right pointer till the string length. whenever we see a repeating character we have to remove that character move the left pointer and start again counting till right-left+1. size./

Time complexity:- O(N)
Space complexity:- O(N) (since we are storing character in the set).